backend/requirements.txt
Django==3.1.1
psycopg2==2.8.6
django-utils-six==2.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
django-cors-headers==3.5.0
python-dotenv==0.15.0
